lesson_3/common/utils.py
import json

from common.variables import ENCODING, MAX_PACKAGE_LENGTH


def send_message(sock, message):  # отправка сообщения
    js_message = json.dumps(message)
    encoded_message = js_message.encode(ENCODING)
    sock.send(encoded_message)


def get_message(client):  # прием сообщения
    encoded_response = client.recv(MAX_PACKAGE_LENGTH)
    if isinstance(encoded_response, bytes):
        json_response = encoded_response.decode(ENCODING)
        response = json.loads(json_response)
        if isinstance(response, dict):
            return response
        raise ValueError
    raise ValueError
